Move SubstrateGCOptions to guest staging
