+++ substratevm/src/com.oracle.svm.guest.staging/src/com/oracle/svm/guest/staging/SubstrateGCOptions.java
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.oracle.svm.core;
+package com.oracle.svm.guest.staging;
 
 import static com.oracle.svm.guest.staging.option.RuntimeOptionKey.RuntimeOptionKeyFlag.Immutable;
 import static com.oracle.svm.guest.staging.option.RuntimeOptionKey.RuntimeOptionKeyFlag.IsolateCreationOnly;
@@ -30,13 +30,13 @@ import static com.oracle.svm.guest.staging.option.RuntimeOptionKey.RuntimeOption
 import static com.oracle.svm.shared.option.HostedOptionKey.HostedOptionKeyFlag.DoNotPassToNativeGC;
 
 import org.graalvm.collections.EconomicMap;
-import com.oracle.svm.guest.staging.GuestStagingDependencyBridge;
+
 import com.oracle.svm.guest.staging.option.NotifyGCRuntimeOptionKey;
-import com.oracle.svm.shared.util.DuplicatedInNativeCode;
-import com.oracle.svm.core.util.UserError;
 import com.oracle.svm.guest.staging.option.RuntimeOptionKey;
+import com.oracle.svm.guest.staging.util.UserError;
 import com.oracle.svm.shared.option.HostedOptionKey;
 import com.oracle.svm.shared.option.SubstrateOptionsParser;
+import com.oracle.svm.shared.util.DuplicatedInNativeCode;
 import com.oracle.svm.shared.util.SubstrateUtil;
 
 import jdk.graal.compiler.options.Option;
@@ -116,8 +116,7 @@ public class SubstrateGCOptions {
     public static final HostedOptionKey<TLABPolicy> TLABUsagePolicy = new HostedOptionKey<>(TLABPolicy.Auto, SubstrateGCOptions::verifyTLABUsagePolicy, DoNotPassToNativeGC);
 
     @Option(help = "Determines whether to outline write barrier code to a separate function, trading reduced image size for (potentially) worse performance.", type = OptionType.Expert) //
-    public static final HostedOptionKey<OutlineWriteBarriers> WriteBarrierOutlining = new HostedOptionKey<>(OutlineWriteBarriers.Auto, SubstrateGCOptions::verifyWriteBarrierOutlining,
-                    DoNotPassToNativeGC);
+    public static final HostedOptionKey<OutlineWriteBarriers> WriteBarrierOutlining = new HostedOptionKey<>(OutlineWriteBarriers.Auto, DoNotPassToNativeGC);
 
     @Option(help = "Dynamically resize TLAB size for threads.", type = OptionType.Expert)//
     public static final RuntimeOptionKey<Boolean> ResizeTLAB = new RuntimeOptionKey<>(true, IsolateCreationOnly);
@@ -132,10 +131,10 @@ public class SubstrateGCOptions {
         }
     }
 
-    private static void verifyWriteBarrierOutlining(HostedOptionKey<OutlineWriteBarriers> key) {
-        OutlineWriteBarriers value = key.getValue();
-        if (SubstrateOptions.useG1GC() && value == OutlineWriteBarriers.YoungOnly) {
-            throw UserError.invalidOptionValue(key, value.name(), "This option value is not supported with the G1 garbage collector");
+    public static void validateWriteBarrierOutlining(boolean useG1GC) {
+        OutlineWriteBarriers value = WriteBarrierOutlining.getValue();
+        if (useG1GC && value == OutlineWriteBarriers.YoungOnly) {
+            throw UserError.invalidOptionValue(WriteBarrierOutlining, value.name(), "This option value is not supported with the G1 garbage collector");
         }
     }
